+++ .gitignore
@@ -15,3 +15,4 @@ run_commnad.txt
 data/Im2GPS3k/im2gps3ktest.zip
 slurm-test.out
 run_sparse_rs.sh
+results/
